+++ src/Drivers/Typo3EnvDriver.php
@@ -17,16 +17,16 @@ class Typo3EnvDriver
      * @throws \Exception
      * @internal param null $params
      */
-    public function getDatabaseConfig($dbMappingFields = null, $absolutePathWithConfig = null)
+    public function getDatabaseConfig($dbMappingFields = null, $absolutePathWithConfig = null): array
     {
-        $absolutePathWithConfig = null === $absolutePathWithConfig ? getcwd() : $absolutePathWithConfig;
+        $absolutePathWithConfig = $absolutePathWithConfig ?? getcwd();
         $absolutePathWithConfig = rtrim($absolutePathWithConfig, DIRECTORY_SEPARATOR);
         $dbSettings = [];
         $this->createEnvFileIfDoesNotExist($absolutePathWithConfig . '/.env');
         if (file_exists($absolutePathWithConfig . '/.env')) {
             (new Dotenv())->load($absolutePathWithConfig . '/.env');
             foreach ($dbMappingFields as $key => $dbMappingField) {
-                $dbSettings[$key] = getenv($dbMappingField);
+                $dbSettings[$key] = $this->getenv($dbMappingField);
             }
         } else {
             throw new \Exception('Missing file "' . $absolutePathWithConfig . '"/.env.');
@@ -41,11 +41,11 @@ class Typo3EnvDriver
     public function createEnvFileIfDoesNotExist($absolutePathWithConfig)
     {
         if (!file_exists($absolutePathWithConfig)) {
-            $host = getenv('MYSQL_HOST');
-            $port = getenv('MYSQL_HOST_PORT') ? getenv('MYSQL_HOST_PORT') : 3306;
-            $username = getenv('MYSQL_USER');
-            $password = getenv('MYSQL_PASSWORD');
-            $databaseName = empty(getenv('MYSQL_DATABASE')) ? '' : getenv('MYSQL_DATABASE');
+            $host = $this->getenv('MYSQL_HOST');
+            $port = $this->getenv('MYSQL_HOST_PORT') ?: 3306;
+            $username = $this->getenv('MYSQL_USER');
+            $password = $this->getenv('MYSQL_PASSWORD');
+            $databaseName = empty($this->getenv('MYSQL_DATABASE')) ? '' : $this->getenv('MYSQL_DATABASE');
             if (empty($databaseName)) {
                 exec('git config --get remote.origin.url', $output);
                 preg_match('/\/([a-zA-Z0-9-_\.]+)\.git/', $output[0], $match);
@@ -54,28 +54,28 @@ class Typo3EnvDriver
                 }
                 $databaseBaseName = 'typo3_' . $match[1];
                 $databaseBaseName = str_replace(['.', '-'], ['_', '_'], $databaseBaseName);
-                if (!empty(getenv('MYSQL_USER')) && !empty(getenv('MYSQL_PASSWORD'))) {
-                    $databaseName = $this->tryToCreateDatabaseIfNotExists($host, $port, getenv('MYSQL_USER'),
-                        getenv('MYSQL_PASSWORD'), $databaseBaseName);
+                if (!empty($this->getenv('MYSQL_USER')) && !empty($this->getenv('MYSQL_PASSWORD'))) {
+                    $databaseName = $this->tryToCreateDatabaseIfNotExists($host, $port, $this->getenv('MYSQL_USER'),
+                        $this->getenv('MYSQL_PASSWORD'), $databaseBaseName);
                 }
-                if (empty($databaseName) && !empty(getenv('MYSQL_ROOT_USER')) && !empty(getenv('MYSQL_ROOT_PASSWORD'))) {
-                    $databaseName = $this->tryToCreateDatabaseIfNotExists($host, $port, getenv('MYSQL_ROOT_USER'),
-                        getenv('MYSQL_ROOT_PASSWORD'), $databaseBaseName);
-                    $username = getenv('MYSQL_ROOT_USER');
-                    $password = getenv('MYSQL_ROOT_PASSWORD');
+                if (empty($databaseName) && !empty($this->getenv('MYSQL_ROOT_USER')) && !empty($this->getenv('MYSQL_ROOT_PASSWORD'))) {
+                    $databaseName = $this->tryToCreateDatabaseIfNotExists($host, $port, $this->getenv('MYSQL_ROOT_USER'),
+                        $this->getenv('MYSQL_ROOT_PASSWORD'), $databaseBaseName);
+                    $username = $this->getenv('MYSQL_ROOT_USER');
+                    $password = $this->getenv('MYSQL_ROOT_PASSWORD');
                 }
             }
             $envDist = file_get_contents(dirname($absolutePathWithConfig) . '/.env.dist');
             $replace = [
-                'TYPO3_CONTEXT' => getenv('TYPO3_CONTEXT'),
-                'INSTANCE' => getenv('INSTANCE'),
+                'TYPO3_CONTEXT' => $this->getenv('TYPO3_CONTEXT'),
+                'INSTANCE' => $this->getenv('INSTANCE'),
                 'MYSQL_DATABASE' => $databaseName,
                 'MYSQL_USER' => $username,
                 'MYSQL_PASSWORD' => $password,
                 'MYSQL_HOST' => $host,
                 'MYSQL_HOST_PORT' => $port,
-                'IMAGE_MAGICK_PATH' => getenv('IMAGE_MAGICK_PATH'),
-                'IMAGE_MAGICK_COLORSPACE' => getenv('IMAGE_MAGICK_COLORSPACE')
+                'IMAGE_MAGICK_PATH' => $this->getenv('IMAGE_MAGICK_PATH'),
+                'IMAGE_MAGICK_COLORSPACE' => $this->getenv('IMAGE_MAGICK_COLORSPACE')
             ];
             $envDistMarkersReplaced = preg_replace_callback(
                 "/\\{([A-Za-z0-9_:]+)\\}/",
@@ -88,7 +88,7 @@ class Typo3EnvDriver
         }
     }
 
-    private function tryToCreateDatabaseIfNotExists($host, $port, $user, $password, $databaseBaseName)
+    private function tryToCreateDatabaseIfNotExists($host, $port, $user, $password, $databaseBaseName): string
     {
         $databaseName = null;
         $mysqli = new \mysqli($host, $user, $password, '', $port);
@@ -106,5 +106,10 @@ class Typo3EnvDriver
             }
         }
         return $databaseName;
+    }
+
+    private function getenv($env)
+    {
+        return $_ENV[$env] ?? null;
     }
 }