+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 ~~~~~~
 
 1) [BUGFIX] Fix compatibility with symfony/dotenv 5.0 which do not use getenv() by default.
+2) [TASK] Add ddev config.
 
 14.1.0
 ~~~~~~